Hide native toggle if sufficient balance
