+++ ui/pages/confirmations/components/confirm/info/shared/gas-fee-token-list-item/gas-fee-token-list-item.test.tsx
@@ -3,14 +3,8 @@ import configureStore from '../../../../../../../store/store';
 import { getMockConfirmStateForTransaction } from '../../../../../../../../test/data/confirmations/helper';
 import { genUnapprovedContractInteractionConfirmation } from '../../../../../../../../test/data/confirmations/contract-interaction';
 import { renderWithConfirmContextProvider } from '../../../../../../../../test/lib/confirmations/render-helpers';
-import { useInsufficientBalanceAlerts } from '../../../../../hooks/alerts/transactions/useInsufficientBalanceAlerts';
-import { Severity } from '../../../../../../../helpers/constants/design-system';
 import { GAS_FEE_TOKEN_MOCK } from '../../../../../../../../test/data/confirmations/gas';
 import { GasFeeTokenListItem } from './gas-fee-token-list-item';
-
-jest.mock(
-  '../../../../../hooks/alerts/transactions/useInsufficientBalanceAlerts',
-);
 
 const FROM_MOCK = '0x0dcd5d886577d5081b0c52e242ef29e70be3e7bc';
 
@@ -32,13 +26,8 @@ const store = configureStore(
 );
 
 describe('GasFeeTokenListItem', () => {
-  const useInsufficientBalanceAlertsMock = jest.mocked(
-    useInsufficientBalanceAlerts,
-  );
-
   beforeEach(() => {
     jest.resetAllMocks();
-    useInsufficientBalanceAlertsMock.mockReturnValue([]);
   });
 
   it('renders fiat amount', () => {
@@ -68,6 +57,15 @@ describe('GasFeeTokenListItem', () => {
     expect(result.getByText('1 TEST')).toBeInTheDocument();
   });
 
+  it('renders warning indicator if warning', () => {
+    const result = renderWithConfirmContextProvider(
+      <GasFeeTokenListItem tokenAddress={undefined} warning="Test Warning" />,
+      store,
+    );
+
+    expect(result.getByText('Test Warning')).toBeInTheDocument();
+  });
+
   describe('with no token address', () => {
     it('renders fiat amount', () => {
       const result = renderWithConfirmContextProvider(
@@ -94,23 +92,6 @@ describe('GasFeeTokenListItem', () => {
       );
 
       expect(result.getByText('0.000066 ETH')).toBeInTheDocument();
-    });
-
-    it('renders insufficient balance indicator', () => {
-      useInsufficientBalanceAlertsMock.mockReturnValue([
-        {
-          content: 'Insufficient balance',
-          key: 'insufficientBalance',
-          severity: Severity.Danger,
-        },
-      ]);
-
-      const result = renderWithConfirmContextProvider(
-        <GasFeeTokenListItem tokenAddress={undefined} />,
-        store,
-      );
-
-      expect(result.getByText('Insufficient funds')).toBeInTheDocument();
     });
   });
 });
