+++ ui/pages/confirmations/components/confirm/info/shared/gas-fee-token-list-item/gas-fee-token-list-item.tsx
@@ -28,29 +28,25 @@ import {
   NATIVE_TOKEN_ADDRESS,
   useGasFeeToken,
 } from '../../hooks/useGasFeeToken';
-import { useInsufficientBalanceAlerts } from '../../../../../hooks/alerts/transactions/useInsufficientBalanceAlerts';
 import { getCurrentCurrency } from '../../../../../../../ducks/metamask/metamask';
 import { GasFeeTokenIcon, GasFeeTokenIconSize } from '../gas-fee-token-icon';
 
 export type GasFeeTokenListItemProps = {
-  tokenAddress?: Hex;
   isSelected?: boolean;
   onClick?: (token: GasFeeToken) => void;
+  tokenAddress?: Hex;
+  warning?: string;
 };
 
 export function GasFeeTokenListItem({
-  tokenAddress,
   isSelected,
   onClick,
+  tokenAddress,
+  warning,
 }: GasFeeTokenListItemProps) {
   const t = useI18nContext();
   const gasFeeToken = useGasFeeToken({ tokenAddress });
   const currentCurrency = useSelector(getCurrentCurrency);
-
-  const hasInsufficientNative =
-    Boolean(useInsufficientBalanceAlerts({ ignoreGasFeeToken: true }).length) &&
-    !tokenAddress;
-
   const isFutureNative = tokenAddress === NATIVE_TOKEN_ADDRESS;
 
   if (!gasFeeToken) {
@@ -76,7 +72,7 @@ export function GasFeeTokenListItem({
       } ${balanceFiat} ${currentCurrency.toUpperCase()}`}
       rightPrimary={amountFiat}
       rightSecondary={`${amountFormatted} ${symbol}`}
-      warning={hasInsufficientNative && <InsufficientBalanceIndicator />}
+      warning={warning && <WarningIndicator text={warning} />}
       onClick={() => onClick?.(gasFeeToken)}
     />
   );
@@ -168,9 +164,7 @@ function ListItem({
   );
 }
 
-function InsufficientBalanceIndicator() {
-  const t = useI18nContext();
-
+function WarningIndicator({ text }: { text: string }) {
   return (
     <Box
       display={Display.Flex}
@@ -187,7 +181,7 @@ function InsufficientBalanceIndicator() {
         color={IconColor.iconMuted}
       />
       <Text variant={TextVariant.bodyXsMedium} color={TextColor.textMuted}>
-        {t('confirmGasFeeTokenInsufficientBalance')}
+        {text}
       </Text>
     </Box>
   );
